fix: resolve frontend lib build files conflict with app build files.
When the frontend build command was run, it saved the build files in the dist folder, the build:lib command also saved the library build files there.
When the frontend build command is run, this removed the lib files in the dist folder which is needed for the pluton-pro frontend's local import for development purpose.