feat: add windows installer update mechanism.
This changes creates the ability to run the install process unlimited time. The installer now detects if the config file exist to detect whether the app is already installed or not.

If its already installed, it stops the service, removes it, does not create the config.json file, copies the latest files, and starts the service.